+++ Menu-o-Matic/Menu-o-Matic.mm
@@ -3,9 +3,9 @@
 <node TEXT="Menu-o-Matic" ID="ID_696401721" LINK="https://github.com/EdoFro/Freeplane_Menu-o-Matic" BACKGROUND_COLOR="#97c7dc">
 <font SIZE="16" BOLD="true" ITALIC="true"/>
 <hook NAME="AutomaticEdgeColor" COUNTER="17" RULE="ON_BRANCH_CREATION"/>
-<attribute_layout NAME_WIDTH="104.25 pt" VALUE_WIDTH="233.24999 pt"/>
+<attribute_layout NAME_WIDTH="112.5 pt" VALUE_WIDTH="266.99999 pt"/>
 <attribute NAME="name" VALUE="menuOMatic"/>
-<attribute NAME="version" VALUE="v0.1.7"/>
+<attribute NAME="version" VALUE="v0.1.8"/>
 <attribute NAME="author" VALUE="EdoFro"/>
 <attribute NAME="freeplaneVersionFrom" VALUE="v1.9.0"/>
 <attribute NAME="freeplaneVersionTo" VALUE=""/>
@@ -260,6 +260,12 @@
 <node TEXT="created changePackParameters command" ID="ID_321714507"/>
 <node TEXT="insertMenuOMaticPackage updated" ID="ID_1959406335"/>
 </node>
+<node TEXT="v0.1.8" ID="ID_49387614">
+<node TEXT="init script opens Menu o Matic toolbar" ID="ID_1115388473"/>
+<node TEXT="Added preference to show/hide Menu-o-Matic toolbar buttons labels" ID="ID_1223570050"/>
+<node TEXT="Added menu separators" ID="ID_261454163"/>
+<node TEXT="Toolbar title can be modified in dialog" ID="ID_419612997"/>
+</node>
 </node>
 <node TEXT="license" POSITION="top_or_left" ID="ID_492801974">
 <edge COLOR="#00ff00"/>
@@ -395,7 +401,8 @@
 </richcontent>
 <attribute_layout NAME_WIDTH="109.5 pt" VALUE_WIDTH="51.75 pt"/>
 <attribute NAME="runOnStartingFreeplane" VALUE="boolean"/>
-<node TEXT="&lt;?xml version=&quot;1.0&quot; encoding=&quot;UTF-8&quot;?&gt;&#xa;&lt;preferences_structure&gt;&#xa;         &lt;tabbed_pane&gt;&#xa;                  &lt;tab name=&quot;plugins&quot;&gt;&#xa;                            &lt;separator name = &quot;${name}&quot;&gt;&#xa;                                    &lt;boolean name = &quot;${name}_runOnStartingFreeplane&quot;/&gt;&#xa;                           &lt;/separator&gt;&#xa;                  &lt;/tab&gt;&#xa;         &lt;/tabbed_pane&gt;&#xa;&lt;/preferences_structure&gt;" ID="ID_1133880285" MAX_WIDTH="20 cm"/>
+<attribute NAME="showMenuOMaticMenuLabels" VALUE="boolean"/>
+<node TEXT="&lt;?xml version=&quot;1.0&quot; encoding=&quot;UTF-8&quot;?&gt;&#xa;&lt;preferences_structure&gt;&#xa;         &lt;tabbed_pane&gt;&#xa;                  &lt;tab name=&quot;plugins&quot;&gt;&#xa;                            &lt;separator name = &quot;${name}&quot;&gt;&#xa;                                    &lt;boolean name = &quot;${name}_runOnStartingFreeplane&quot;/&gt;&#xa;                                    &lt;boolean name = &quot;${name}_showMenuOMaticMenuLabels&quot;/&gt;&#xa;                           &lt;/separator&gt;&#xa;                  &lt;/tab&gt;&#xa;         &lt;/tabbed_pane&gt;&#xa;&lt;/preferences_structure&gt;" ID="ID_1133880285" MAX_WIDTH="20 cm"/>
 </node>
 <node TEXT="default.properties" POSITION="top_or_left" ID="ID_252406354">
 <edge COLOR="#00ffff"/>
@@ -419,8 +426,9 @@
   </body>
 </html>
 </richcontent>
-<attribute_layout NAME_WIDTH="144 pt" VALUE_WIDTH="37.5 pt"/>
+<attribute_layout NAME_WIDTH="182.24999 pt" VALUE_WIDTH="39.75 pt"/>
 <attribute NAME="${name}_runOnStartingFreeplane" VALUE="true"/>
+<attribute NAME="${name}_showMenuOMaticMenuLabels" VALUE="true"/>
 </node>
 <node TEXT="translations" POSITION="top_or_left" ID="ID_971883114">
 <edge COLOR="#7c0000"/>
@@ -456,10 +464,11 @@
   </body>
 </html>
 </richcontent>
-<node TEXT="en" ID="ID_612862351">
-<attribute_layout NAME_WIDTH="212.24999 pt" VALUE_WIDTH="411.74999 pt"/>
+<node TEXT="en" ID="ID_43497759">
+<attribute_layout NAME_WIDTH="262.49999 pt" VALUE_WIDTH="456.74999 pt"/>
 <attribute NAME="addons.${name}.lauchDialogFromNodes" VALUE="launch menu dialog directly"/>
 <attribute NAME="addons.${name}.1launchMenuOMaticDialog" VALUE="launch Menu-o-Matic dialog"/>
+<attribute NAME="OptionPanel.${name}_showMenuOMaticMenuLabels" VALUE="show buttons labels in Menu-o-Matic toolbar"/>
 <attribute NAME="addons.${name}.launchCustomDialog" VALUE="launch menu dialog from package"/>
 <attribute NAME="addons.${name}.listCustomMenusFromMap" VALUE="list menu packages from map"/>
 <attribute NAME="addons.${name}" VALUE="Menu-o-Matic"/>
@@ -470,6 +479,7 @@
 <attribute NAME="addons.${name}.showLastUsedMenu" VALUE="show most recently used menu"/>
 <attribute NAME="OptionPanel.separator.${name}" VALUE="Menu-o-Matic"/>
 <attribute NAME="OptionPanel.${name}_runOnStartingFreeplane.tooltip" VALUE="Automatically opens all Custom Menus marked as AutoLaunch from all opened mindmaps on Freeplane start"/>
+<attribute NAME="OptionPanel.${name}_showMenuOMaticMenuLabels.tooltip" VALUE="Defines if MoM toolbar shows   icons and labels   or   icons only"/>
 <attribute NAME="addons.${name}.getMenuCommand" VALUE="menu item under mouse as new node"/>
 <attribute NAME="addons.${name}.insertMenuOMaticPackage" VALUE="Insert Menu-o-Matic package node"/>
 <attribute NAME="addons.${name}.listCustomDialogs" VALUE="list opened menu dialogs"/>