+++ Menu-o-Matic/src/main/groovy/LaunchDialog.groovy
@@ -1,39 +1,38 @@
 package edofro.menuomatic
 
-import edofro.menuomatic.LaunchTabPane
-import edofro.menuomatic.MoMToolbar
-import groovy.swing.SwingBuilder
-import org.freeplane.api.MindMap
-import org.freeplane.core.ui.components.ToolbarLayout
-import org.freeplane.features.map.MapModel
-import org.freeplane.plugin.script.proxy.MapProxy
-
-
-import javax.swing.Icon
-import javax.swing.JButton
-import javax.swing.SwingConstants
 import java.awt.Color
+import java.awt.Dimension
 import java.awt.GridLayout
 import java.awt.Insets
-import java.awt.Dimension
 import java.awt.event.WindowFocusListener
+import javax.swing.BoxLayout
+import javax.swing.Icon
+import javax.swing.JButton
+import javax.swing.JSeparator
+import javax.swing.SwingConstants
 
-
+import groovy.swing.SwingBuilder
 import groovy.time.TimeCategory
 import groovy.time.TimeDuration
 
-
-import edofro.menuomatic.PackMenu                   as PM
-import edofro.menuomatic.DialogKeyboardNavigation   as DKBN
-    
+import org.freeplane.api.MindMap
+import org.freeplane.core.ui.components.ToolbarLayout
 import org.freeplane.core.ui.components.UITools     as ui
-import org.freeplane.core.util.TextUtils            as textUtils
 import org.freeplane.core.util.MenuUtils            as menuUtils
+import org.freeplane.core.util.TextUtils            as textUtils
+import org.freeplane.features.map.MapModel
+import org.freeplane.plugin.script.proxy.MapProxy
 import org.freeplane.plugin.script.proxy.ScriptUtils
+
+import edofro.menuomatic.DialogKeyboardNavigation   as DKBN
+import edofro.menuomatic.LaunchTabPane
+import edofro.menuomatic.MoMToolbar
+import edofro.menuomatic.PackMenu                   as PM
+
 
 
 class LaunchDialog{
-//region: properties
+//region properties
     static final String defaultIcon     = 'IconAction.emoji-1F7EB'
     static final String dialogStr       = '_MoM_'
     static final int    maxButtonsHoriz = 6
@@ -42,6 +41,7 @@ class LaunchDialog{
     static final c                          = ScriptUtils.c()
     static final tb                         = PM.TB
     static final scriptStr                  = PM.scriptStr
+    static final separatorStr               = PM.separatorStr
 
     static PM.MenuData md
     static Dimension prefDimension
@@ -52,7 +52,7 @@ class LaunchDialog{
     static final int WITHOUT_EXEC_RESTRICTION    = 0b0100
     static final int WITHOUT_NETWORK_RESTRICTION = 0b1000
 
-//endregion:
+//endregion
 
 
     // region managing dialogs
@@ -211,7 +211,8 @@ class LaunchDialog{
     }
 
     def static setUpToolbar(tb) {
-        def theLayout = md.showLabels?new GridLayout(0,1): ToolbarLayout.vertical()
+        //def theLayout = md.showLabels?new GridLayout(0,1): ToolbarLayout.vertical()
+        def theLayout = md.showLabels?new BoxLayout(tb, BoxLayout.PAGE_AXIS): ToolbarLayout.vertical()  // TODO : BoxLayout.PAGE_AXIS
         tb.setLayout(theLayout)
         tb.setFloatable(true)
         tb.margin = new Insets(0, 0, 5, 0)
@@ -228,7 +229,7 @@ class LaunchDialog{
 
     // endregion
 
-    //region creating panel
+    // region creating panel
 
     def static getCols(){
         md.showLabels?1:buttonCols(md.actions.size(),maxButtonsHoriz)
@@ -280,6 +281,8 @@ class LaunchDialog{
     def static creaBoton(acc, i){
         if(acc.startsWith(scriptStr)){
             return creaBotonDesdeScript(acc,i)
+        } else if(acc==separatorStr){
+            return new JSeparator(SwingConstants.HORIZONTAL) //SI NO FUNCIONA. VER 
         } else {
             return creaBotonDesdeUI(acc, i)
         }