+++ Menu-o-Matic/src/main/groovy/PackMenu.groovy
@@ -20,12 +20,14 @@ import groovy.transform.MapConstructor
 
 class PackMenu{
 
-    // region: properties
+    // region properties
 
     static final String   scriptStr    = '_script'
+    static final String   separatorStr = '_separator'
     // static final Boolean  scriptInNote = true
     
     static final String MoMDialogTitle = 'menu-o-matic'
+    static final String nameMsg = 'menu\'s name:'
     static final String iconLabelMsg = 'Dialog must show:'
     static final String[] optionsD1    = ['icons only', 'labels only','icons and labels','cancel']
     static final String   titleD2      = 'menu-o-matic'
@@ -35,7 +37,7 @@ class PackMenu{
     //static final String[] optionsD3    = ['return to mindmap','stay in menu']  //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sense again.
     static final String   MoM_TAB_NAME = 'MoM'
     static final String   textLengthMsg = 'Maximal label text length:'
-    static final String   tabNameMsg    = 'Include toolbar to tab:'
+    static final String   tabNameMsg    = 'Place toolbar into tab:'
     static final String   autoLaunchMsg = 'Auto launch toolbar'
     static final String   permissionsMsg = 'Scripts permissions'
     static final String   readPermission  = 'WITHOUT_READ_RESTRICTION'
@@ -62,7 +64,7 @@ class PackMenu{
     ]
 
 
-    // endregion: properties
+    // endregion properties
 
     @MapConstructor
     static class MenuData{
@@ -91,7 +93,7 @@ class PackMenu{
                 this.showIcons      = nodoMenu[TB.showIcons ].bool
                 this.showLabels     = nodoMenu[TB.showLabels].bool
                 //this.focusMap       = nodoMenu[TB.focusMap  ].bool //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sence again.
-                this.title          = nodoMenu[TB.title]?nodoMenu[TB.title].toString():null
+                this.title          = nodoMenu[TB.title].string?:null
                 def scriptList = []
                 nodoMenu.attributes.names.findAll{it.startsWith(scriptStr)}.each{
                     scriptList << [it,nodoMenu[it]]
@@ -106,6 +108,7 @@ class PackMenu{
 //            }
         }
 
+        //PackMenu$MenuData(String, org.freeplane.plugin.script.proxy.ProxyUtils$1, Boolean, Boolean, Boolean, Long, String, Boolean, null)
         public MenuData(String title, nAcciones, boolean showIcons, boolean showLabels,
                         boolean focusMap, int maxTextLength, String tabName, boolean autoLaunch, String permissions){
             iScript      = 0
@@ -163,18 +166,19 @@ class PackMenu{
     static final c = ScriptUtils.c()
 
 
-    // region: primary methods
+    // region primary methods
 
     def static getMDfromNodes(nodoBase, boolean useDetails = true){
         //get info from child nodes
-        def nAcciones  = nodoBase.find{it.link?.uri?.scheme == 'menuitem' || WSE.isGroovyNode(it)}
+        def nAcciones  = nodoBase.find{it.link?.uri?.scheme == 'menuitem' || WSE.isGroovyNode(it) || isSeparatorNode(it)}
         def hasScripts = nAcciones.any{WSE.isGroovyNode(it)}
 
         //get info from node
-        def title = nodoBase.text
+        //def title = nodoBase.text
 
         //get info from node/dialog
-        def maxTextLength = nodoBase[TB.maxTextLength].num
+        def menuName      = nodoBase[TB.title].string?:nodoBase.text
+        def maxTextLength = nodoBase[TB.maxTextLength].num?.toInteger()
         def tabName       = nodoBase[TB.tabName].string
         def autoLaunch    = nodoBase[TB.autoLaunch]//.bool
         def showIcons     = nodoBase[TB.showIcons]//.bool
@@ -182,9 +186,10 @@ class PackMenu{
         def focusToMap    = nodoBase[TB.focusMap]//.bool
         def permissions   = nodoBase[TB.permissions].string
         def forceDialog = false
-        (maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions) = getConfirmedInfo(forceDialog, hasScripts, maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions)
-        if(maxTextLength && tabName && autoLaunch!=null && showIcons!=null && showLabels!=null && focusToMap!=null) {
-            return [new MenuData(nodoBase.text, nAcciones, showIcons, showLabels, focusToMap, maxTextLength, tabName, autoLaunch, permissions), "\n  - resp\n  - resp2" ]
+        (menuName, maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions) = getConfirmedInfo(forceDialog, hasScripts, menuName, maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions)
+        menuName?=nodoBase[TB.title].string?:nodoBase.text
+        if(menuName && maxTextLength && tabName && autoLaunch!=null && showIcons!=null && showLabels!=null && focusToMap!=null) {
+            return [new MenuData(menuName, nAcciones, showIcons, showLabels, focusToMap, maxTextLength, tabName, autoLaunch, permissions), "\n  - resp\n  - resp2" ]
         } else {
             return [null, 'resp']
         }
@@ -231,7 +236,8 @@ class PackMenu{
         def fgColor        = nodoMenu.style.textColorSet?nodoMenu.style.textColorCode:null
         if(fgColor)
             nBase.style.textColorCode = fgColor
-        nBase[TB.showIcons] = md.showIcons
+        nBase[TB.title]      = md.title
+        nBase[TB.showIcons]  = md.showIcons
         nBase[TB.showLabels] = md.showLabels
         //nBase[TB.focusMap] = md.focusMap //focus to map is currently don't needed, I will let the code here just in case in a future version it makes sence again.
         nBase[TB.maxTextLength] = md.maxTextLength
@@ -248,6 +254,8 @@ class PackMenu{
                     nodo['script1'] = nodoMenu[acc].toString()
                 }
                 WSE.setExtension(nodo, 'groovy')
+            } else if (acc==separatorStr){
+                //nodo.text = '---' // it seems it isn't necessary
             } else {
                 nodo.link.text = "menuitem:_${acc}"
             }
@@ -276,12 +284,13 @@ class PackMenu{
         return "$title  ($iconLabel, $focus)".toString()
     }
 
-    def static getConfirmedInfo(forceDialog, hasScripts, maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions) {
+    def static getConfirmedInfo(forceDialog, hasScripts, menuName, maxTextLength, tabName, autoLaunch, showIcons, showLabels, focusToMap, permissions) {
         //if all info is ready discard inputDialog
-        if(!forceDialog && maxTextLength && tabName && autoLaunch!=null && showIcons!=null && showLabels!=null && focusToMap!=null && (!hasScripts || permissions)){
-            return [maxTextLength, tabName, autoLaunch.bool, showIcons.bool, showLabels.bool, focusToMap.bool, permissions]
+        if(!forceDialog && maxTextLength && tabName && menuName && autoLaunch!=null && showIcons!=null && showLabels!=null && focusToMap!=null && (!hasScripts || permissions)){
+            return [menuName, maxTextLength, tabName, autoLaunch.bool, showIcons.bool, showLabels.bool, focusToMap.bool, permissions]
         }
         //create input objects
+        JTextField menuNameField = new JTextField(menuName)
         JTextField textLengthField = new JTextField(maxTextLength?.toString()?:maxTextLen.toString())
         JTextField tabNameField = new JTextField(tabName?:MoM_TAB_NAME)
         JCheckBox autoLaunchCheckBox = new JCheckBox(autoLaunchMsg, autoLaunch?autoLaunch.bool:false)
@@ -297,6 +306,8 @@ class PackMenu{
         JCheckBox exePermissionCheckBox   = hasScripts? new JCheckBox(exePermission,   (binPermissions & 0b1000)>0): null
         //create group component
         final JComponent[] inputs = new JComponent[] {
+                new JLabel(nameMsg),
+                menuNameField,
                 new JLabel(iconLabelMsg),
                 iconsLabelsComboBox,
                 new JLabel(textLengthMsg),
@@ -316,6 +327,7 @@ class PackMenu{
         if (result == JOptionPane.OK_OPTION) {
             def permiString = hasScripts? Integer.toBinaryString((readPermissionCheckBox.selected?0b0001:0)+(writePermissionCheckBox.selected?0b0010:0)+(netPermissionCheckBox.selected?0b0100:0)+(exePermissionCheckBox.selected?0b1000:0)) : null
             return [
+                    menuNameField.text,
                     textLengthField.text?.isInteger()?textLengthField.text.toInteger():maxTextLen,
                     tabNameField.text?:MoM_TAB_NAME,
                     autoLaunchCheckBox.isSelected(),
@@ -325,16 +337,19 @@ class PackMenu{
                     hasScripts? ('0000' + permiString).takeRight(4) : null,
             ]
         } else {
-            return [null, null, null, null, null, null, null]
+            return [null, null, null, null, null, null, null, null]
         }
     }
 
-    // endregion: primary methods
+    // endregion primary methods
 
-    // region: secondary methods
+    // region secondary methods
+    def static isSeparatorNode(n){
+        return n.text.every{ it == '-' }
+    }
 
     def static accion(n){
-        return (n.link?.uri?.scheme == 'menuitem')?n.link.uri.schemeSpecificPart.drop(1):WSE.isGroovyNode(n)?scriptStr + ++iScript :null
+        return (n.link?.uri?.scheme == 'menuitem')?n.link.uri.schemeSpecificPart.drop(1):WSE.isGroovyNode(n)?scriptStr + ++iScript :isSeparatorNode(n)?separatorStr:null
     }
 
     def static icono(n){
@@ -348,5 +363,5 @@ class PackMenu{
         return pane.inputValue
     }
 
-    // endregion: methods
+    // endregion methods
 }